+++ src/libretro.cpp
@@ -328,7 +328,11 @@ bool retro_load_game(const struct retro_game_info *info) {
 	if (gamePath == ".") {
 		gamePath = "main.chai";
 	}
-	return ChaiLove::getInstance()->load(gamePath, info->data);
+	void* data = NULL;
+	if (info != NULL) {
+		data = (void*)info->data;
+	}
+	return ChaiLove::getInstance()->load(gamePath, data);
 }
 
 /**