+++ src/love/script.cpp
@@ -290,7 +290,9 @@ script::script(const std::string& file) {
 	chai.add(fun(&filesystem::createDirectory), "createDirectory");
 	chai.add(fun(&filesystem::isSymlink), "isSymlink");
 	chai.add(fun(&filesystem::isFile), "isFile");
+	chai.add(fun(&filesystem::getWorkingDirectory), "getWorkingDirectory");
 	chai.add(fun(&filesystem::write), "write");
+	chai.add(fun(&filesystem::remove), "remove");
 	chai.add(fun(&filesystem::exists), "exists");
 	chai.add(fun(&filesystem::getSaveDirectory), "getSaveDirectory");
 	chai.add(fun(&filesystem::getInfo), "getInfo");