+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 0.26.2 - Unreleased
+### Fixes
+- Fixed loading ChaiLove without active content
 
 ## 0.26.1 - 2018-09-19
 ### Fixes