+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.26.2 - Unreleased
+## 0.26.2 - 2018-09-24
 ### Features
 - Live updating of core options
 - Mounts `/libretro/core` as the directory where the core was loaded from